Update style.css
This change (--border-tweet-quote) is made to differences border boxes from actual tweets and quote tweets if you use full transparency on --border from settings.
I have no knowledge of javascript, so i am not sure how it is set (maybe in php, idk) to appear as customizable color in the autogenerated list of the "Advanced colors" in settings, so i don't know if needs any change in other document or it just works.

Also borders between tweets disappear making it more messy and this would need a solution a lot more extended in my opinion (maybe some day) and add other things like change the width of pixels and the style. At least this should work for now to see quote tweets.